+++ android/app/src/main/java/com/app/belcobtm/domain/transaction/interactor/trade/CreateBuyTradeUseCase.kt
@@ -20,7 +20,7 @@ class CreateBuyTradeUseCase(private val repository: TransactionRepository) :
     data class Params(
         val coinCode: String,
         val paymentMethod: String,
-        val margin: Int,
+        val margin: Double,
         val minLimit: Long,
         val maxLimit: Long,
         val terms: String
